configure test_service properly
